Add a configurable DC_OFFSET (i.e. which value is "0" for the audio) to the config.
The intention is that the audio range is from -127 to 127, which in unsigned values is 0 to 255. However, due to the nature of electronics, this isn't usually spot on. Some visualisers don't look good if the value we want to be "quiet" is too far away from 0 (especially render_combo_samples_with_beat()).

TODO: estimate this at startup? Or use a differential instead of absolute samples?