adjust samples by the DC_OFFSET for render_combo_samples_with_beat(), otherwise it just looks "very blue".
The intention is that the audio range is from -127 to 127, which in unsigned values is 0 to 255. However, due to the nature of electronics, this isn't usually spot on. Some visualisers don't look good if the value we want to be "quiet" is too far away from 0 (especially render_combo_samples_with_beat()).